Extract Result class
